style: fixed code to meet PEP8 standards.
One line doesn't meet the requirements because it is a url link.  As PEP8 states:
"But most importantly: know when to be inconsistent -- sometimes the style guide just doesn't apply. When in doubt, use your best judgment."

When applying the rule would make the code less readable, even for someone who is used to reading code that follows the rules."